requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
llama-cpp-python>=0.3.0
pydantic>=2.0.0
